+++ ViT-Huge Training/DeepSpeed.py
@@ -121,7 +121,8 @@ def train():
             torch.cuda.synchronize()
             iter_start = time.time()
 
-            images, labels = images.to(local_rank), labels.to(local_rank)
+            # Convert images to half precision
+            images, labels = images.to(local_rank).half(), labels.to(local_rank)
 
             optimizer.zero_grad()
             loss = criterion(model(images), labels)
